Remove 'header' section.
This confuses things more than it helps them, especially when we're trying to reconcile sections in a PE header to those in a BlockGraph and then again to those in an ImageLayout.

This removes special treatment of header blocks with 'kHeaderSectionId' and instead leaves marked with 'kInvalidSectionId'. The same logic applies; a block graph should only ever contain two blocks with section == kInvalidSectionId.

BUG=
TEST=

Review URL: http://codereview.appspot.com/5307068

git-svn-id: http://sawbuck.googlecode.com/svn/trunk@503 15e8cca8-e42c-11de-a347-f34a4f72eb7d