Add full module signatures to RPC call-traces.
This CL adds full module signatures to TraceFileHeader and TraceModuleData (checksum, address, size, time-date stamp) so that modules can be reliably found and validated when playing back a trace file. This CL also modifies the RPC parser to ensure that EXEs also get an entry in the process map, not just DLLs. This ensures that trace events on instrumented EXEs will map to a module.

BUG=
TEST=

Review URL: http://codereview.appspot.com/5375054

git-svn-id: http://sawbuck.googlecode.com/svn/trunk@523 15e8cca8-e42c-11de-a347-f34a4f72eb7d