Remove errant DCHECK when redirecting instrumentation thunks
We have seen examples (albeit arguably malformed) where the same
TLS initializer thunk appears twice, and is therefore rewritten
twice. This was firing a DCHECK that assumed it would only be
redirected once. The second redirect is essentially a NOP and
is therefore harmless, so the DCHECK has been removed.

BUG=
TEST=

Review URL: https://codereview.appspot.com/5556052

git-svn-id: http://sawbuck.googlecode.com/svn/trunk@627 15e8cca8-e42c-11de-a347-f34a4f72eb7d