Remove .egg from .gitignore
Part of the problem with yesterday's P0 fix was that a 'git add *' silently fails to add .egg files, and obviously a 'git status' does not show a stranded .egg as not belonging to the repository.

While our build system still produces stray egg-info directories, it seems to be clear of .egg files after a build all on a fresh checkout.

BUG=
TEST=

Review URL: https://codereview.appspot.com/6013046

git-svn-id: http://sawbuck.googlecode.com/svn/trunk@783 15e8cca8-e42c-11de-a347-f34a4f72eb7d