Initial assembler implementation.
The idea is that the OperandImpl and ImmediateImpl classes can be extended by the user to support a particular reference type.
The AssemblerImpl class should be wrapped in a template class which could provide type safety while delegating encoding to the Impl class.

This isn't quite ready for submission yet, there's still more testing to be done.
The initial instruction implemented is mov in a few varieties.
This is because mov has the greatest variety of encodings, and so should prove the interface and its implementation.

R=rogerm@chromium.org,chrisha@chromium.org

Review URL: https://codereview.appspot.com/6119044

git-svn-id: http://sawbuck.googlecode.com/svn/trunk@792 15e8cca8-e42c-11de-a347-f34a4f72eb7d