Create WritablePdbStream.
This CL creates WritablePdbStream, and augments PdbStream with a 'factory' for getting access to the writable interface, if it is supported. This is not strictly an interface right now, but actually inherits from BufferWriter as a hack. Judicious comments explain the current situation, but eventually we'll want to hoist some proper interfaces here.

BUG=
TEST=

Review URL: https://codereview.appspot.com/6164044

git-svn-id: http://sawbuck.googlecode.com/svn/trunk@796 15e8cca8-e42c-11de-a347-f34a4f72eb7d