Add base address to BlockGraph::Reference.
This CL adds one last piece of information to references, allowing them the full expressive power of FIXUP entries. This will allow us to properly track and fix references when decomposing to basic blocks. This extra information also allowed us to tighten up certain assumptions that were being made. For example, TypedBlock's dereferencing functions only make sense in the context of a direct reference. This allows us to explicitly determine if the reference is direct, and to return failure in the case where it is not.

BUG=
TEST=

Review URL: https://codereview.appspot.com/6256071

git-svn-id: http://sawbuck.googlecode.com/svn/trunk@858 15e8cca8-e42c-11de-a347-f34a4f72eb7d