Refactor Decomposer::OnInstruction.
This CL breaks up OnInstruction into a handful of smaller utility functions. It also changes logging so that known (and frequently occurring) boundary cases are output as VLOG(1), unexpected (but not fatal behaviour) is output as LOG(WARNING), and unrecoverable errors are output as LOG(ERROR). This keeps the output to a reasonable level during normal decomposition. (The new strict disassembly was introducing quite a few new warnings.)

BUG=
TEST=

Review URL: https://codereview.appspot.com/6306087

git-svn-id: http://sawbuck.googlecode.com/svn/trunk@904 15e8cca8-e42c-11de-a347-f34a4f72eb7d