Sample code to resolve the code coverage issue.
In this code I iterate over all the different kind of leaves and for each of them I generate a fake type info stream and I call the Read and the Dump function.
It allow to make sure to cover all the code of the unimplemented kind of leaves. For those who have an implementation this unittest will log an error (but we don't care about that because it's an unittest) and we should relay on the "valid_type_info.pdb_stream" to cover those implementations.

If you think that it can be a good solution to get a better coverage I will clean this code and do the same kind of thing for the symbol record stream.

BUG=
TEST=

Review URL: https://codereview.appspot.com/6343061

git-svn-id: http://sawbuck.googlecode.com/svn/trunk@959 15e8cca8-e42c-11de-a347-f34a4f72eb7d